requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorzero==2.0
gpiozero==1.6.2
idna==2.10
influxdb==5.2.3
pyserial==3.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
six==1.16.0
urllib3==1.26.6

